Create DST & K Data for NFLFastR
Creating fantasy points tables for kickers and DSTs which currently do not exist in "load_player_stats" NFLFastR table. I used Yahoo defensive settings which do not penalize DSTs for points the offense gives up on turnovers. Kicking is pretty standard, 3, 4 and 5 points for 30, 40 and 50 yard field goals respectively. The only thing I know I don't have written in the code is passing/rushing stats for kickers in the event they throw on a fake or whatever.